Added TLS annotations to MariaDB exporter
